Add cron to ensure-lifecyle and check-license
Projects/Repos may not always specify dependencies that are constant over time.
As an example, devcontainer-base-images v0.3 is not constant over time as it may be
patched.
By using a cron job it is possible to detect if there are any changes that must be reflected
in the repository, for example by updating workflows or NOTICE file.

Cron interval identical to interval in
https://github.com/eclipse-velocitas/vehicle-app-cpp-template/blob/main/.github/workflows/check-devcontainer.yml